pyproject.toml
[project]
name = "swe-bench-green-agent"
version = "0.1.0"
description = "A2A green agent that evaluates coding agents on SWE-bench Pro"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
